+++ examples/ex_hdr_images.py
@@ -1,4 +1,8 @@
-from ipaddress import collapse_addresses
+# This example shows how to run ColorVideoVDP from python on HDR images.
+
+# Important: This and other examples should be executed from the main ColorVideoVDP directory:
+# python examples/ex_<...>.py
+
 import os
 import numpy as np
 import matplotlib.pyplot as plt
@@ -8,8 +12,8 @@ import pycvvdp
 
 '''
 Results of current version (for reference):
-Noise - Quality: 9.201 JOD
-Blur - Quality: 8.449 JOD
+Noise - Quality: 9.450 JOD
+Blur - Quality: 8.696 JOD
 '''
 
 I_ref = pycvvdp.load_image_as_array(os.path.join('example_media', 'nancy_church.hdr'))