+++ pycvvdp/utils.py
@@ -3,6 +3,10 @@ import torch
 import numpy as np
 import json
 import torch.nn.functional as Func
+import math
+from functools import cache
+
+from pycvvdp.interp import interp1, interp1q
 #from PIL import Image
 
 from pycvvdp.third_party.loadmat import loadmat
@@ -185,28 +189,45 @@ class PU():
         self.L_max = L_max
 
         if type == 'banding':
-            self.p = [1.063020987, 0.4200327408, 0.1666005322, 0.2817030548, 1.029472678, 1.119265011, 502.1303377]
+            self.p = [1.070275272, 0.4088273932, 0.153224308, 0.2520326168, 1.063512885, 1.14115047, 521.4527484]
         elif type == 'banding_glare':
-            self.p = [234.0235618, 216.9339286, 0.0001091864237, 0.893206924, 0.06733984121, 1.444718567, 567.6315065];
+            self.p = [0.353487901, 0.3734658629, 8.277049286e-05, 0.9062562627, 0.09150303166, 0.9099517204, 596.3148142]
         elif type == 'peaks':
-            self.p = [1.057454135, 0.6234292574, 0.3060331179, 0.3702234502, 1.116868695, 1.109926637, 391.3707005];
+            self.p = [1.043882782, 0.6459495343, 0.3194584211, 0.374025247, 1.114783422, 1.095360363, 384.9217577]
         elif type == 'peaks_glare':
-            self.p = [1.374063733, 0.3160810744, 0.1350497609, 0.510558148, 1.049265455, 1.404963498, 427.3579761];
+            self.p = [816.885024, 1479.463946, 0.001253215609, 0.9329636822, 0.06746643971, 1.573435413, 419.6006374]
         else:
             raise ValueError(f'Unknown type: {type}')
 
         self.peak = self.p[6]*(((self.p[0] + self.p[1]*L_max**self.p[3])/(1 + self.p[2]*L_max**self.p[3]))**self.p[4] - self.p[5])
 
-    def encode(self, Y):
+    def _encode_direct(self, Y):
         '''
         Convert from linear (optical) values Y to encoded (electronic) values V
         '''
         # epsilon = 1e-5
         # if (Y < (self.L_min - epsilon)).any() or (Y > (self.L_max + epsilon)).any():
         #     print( 'Values passed to encode are outside the valid range' )
-
         Y = Y.clip(self.L_min, self.L_max)
-        V = self.p[6]*(((self.p[0] + self.p[1]*Y**self.p[3])/(1 + self.p[2]*Y**self.p[3]))**self.p[4] - self.p[5])
+        Y_p = Y**self.p[3]
+        V = self.p[6]*(((self.p[0] + self.p[1]*Y_p)/(1 + self.p[2]*Y_p))**self.p[4] - self.p[5])
+        return V
+
+    @cache
+    def _get_encode_lut(self, device):
+        Y_lut = torch.linspace(math.log10(self.L_min), math.log10(self.L_max), 2048, device=device)
+        V_lut = self._encode_direct(10**Y_lut)
+        return (Y_lut, V_lut)
+
+    def encode(self, Y):
+        '''
+        Convert from linear (optical) values Y to encoded (electronic) values V
+        '''
+        if True or Y.numel()<100: # Interpolation seems slower than directly computing the values
+            V = self._encode_direct(Y)
+        else:
+            (Y_lut, V_lut) = self._get_encode_lut(Y.device)
+            V = interp1q( Y_lut, V_lut, torch.log10(Y.clamp(self.L_min, self.L_max)) )
         return V
 
     def decode(self, V):