+++ ChangeLog.md
@@ -1,5 +1,31 @@
-# v0.? (?)
+# v0.4.2 (29/09/2024)
+* Files are now sorted after the wildcard expansion
+* Updated PU21 encoding parameters so that they are in sync with those in https://github.com/gfxdisp/pu21/
+* Added: `--dump-channels` for generating videos with intermediate processing stages (debugging and visualization)
+* Added: Support for HLG EOTF (e.g. iPhone HDR video) - thanks to Cosmin Stejerean
+* Added: Processing of videos stored as image frames, described using the C-notation `frame_%04d.png`. New arguments: '--fps' and '--frames'
+* Fixed: A better memory model for estimating how many frames can be processed at once on a GPU. Added '--gpu-mem' argument.
+* Added: 'exposure' field in a display model JSON file.
+* Added: `--result` argument to store results in a CSV file.
+* Fixed: Added examples to README.md and improved documentation.
+* Added: ColorVideoVDP logo.
+
+# v0.4.1 (27/04/2024)
+* Added `--metric dm-preview` for debugging of the display model
+* Added a `loss` method to cvvdp and `examples/ex_adaptive_chroma_subsampling.py`
+* Added "pixels_per_degree" as an optional field in `display_models.json`
+* Platform is printed when running with `--verbose` (for reporting bugs/issues)
+
+# v0.4 (19/01/2024)
+* A new calibration with a small improvement in performance
+* Improved predictions for supra-threshold contrast across color directions
+* Improved masking model
 * Added Matlab wrapper
+* CLI now has --interactive mode to process multiple images/video without restarting PyTorch
+* Fixed distogram generation for images
+* "--quiet" flag now ensures that no warning messages end up in the stdout
+* CSF updated to the latest castleCSF fit
+* added 'luminance' color space to handle luminance-only data
 
 # v0.3 (05/09/2023)
 * Added value range checks when the metric is running on HDR data (to avoid passing relative values)